# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HitOut": {
        "properties": {
          "external_key": {
            "title": "External Key",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "matched_chunks": {
            "items": {
              "$ref": "#/components/schemas/memoryos__api__routes__search__ChunkOut"
            },
            "title": "Matched Chunks",
            "type": "array"
          },
          "memory_id": {
            "format": "uuid",
            "title": "Memory Id",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "memory_id",
          "external_key",
          "source_name",
          "title",
          "kind",
          "occurred_at",
          "score",
          "matched_chunks"
        ],
        "title": "HitOut",
        "type": "object"
      },
...
      "MemoryKind": {
        "enum": [
          "note",
          "document",
          "code",
          "email",
          "commit",
          "meeting",
          "bookmark",
          "other"
        ],
        "title": "MemoryKind",
        "type": "string"
      },
...
      "SearchIn": {
        "properties": {
          "after": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "before": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "ef_search": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ef Search"
          },
          "exact": {
            "default": false,
            "title": "Exact",
            "type": "boolean"
          },
          "include_deleted": {
            "default": false,
            "title": "Include Deleted",
            "type": "boolean"
          },
          "k": {
            "default": 10,
            "title": "K",
            "type": "integer"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "q": {
            "minLength": 1,
            "title": "Q",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "q"
        ],
        "title": "SearchIn",
        "type": "object"
      },
      "SearchOut": {
        "properties": {
          "hits": {
            "items": {
              "$ref": "#/components/schemas/HitOut"
            },
            "title": "Hits",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingOut"
          }
        },
        "required": [
          "query",
          "hits",
          "timing"
        ],
        "title": "SearchOut",
        "type": "object"
      },
...
      "TimingOut": {
        "properties": {
          "embed_ms": {
            "title": "Embed Ms",
            "type": "integer"
          },
          "search_ms": {
            "title": "Search Ms",
            "type": "integer"
          },
          "total_ms": {
            "title": "Total Ms",
            "type": "integer"
          }
        },
        "required": [
          "embed_ms",
          "search_ms",
          "total_ms"
        ],
        "title": "TimingOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "memoryos__api__routes__search__ChunkOut": {
        "properties": {
          "char_end": {
            "title": "Char End",
            "type": "integer"
          },
          "char_start": {
            "title": "Char Start",
            "type": "integer"
          },
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "chunk_id",
          "ordinal",
          "text",
          "score",
          "char_start",
          "char_end"
        ],
        "title": "ChunkOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Memory Intelligence OS",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/search": {
      "get": {
        "operationId": "search_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "k",
            "required": false,
            "schema": {
              "default": 10,
              "title": "K",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MemoryKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "ef_search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ef Search"
            }
          },
          {
            "in": "query",
            "name": "exact",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exact",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "search"
        ]
      },
      "post": {
        "description": "Same thing, for queries too long to sit comfortably in a URL.",
        "operationId": "search_post_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Post",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}